pyproject.toml
[project]
name = "convert-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-core>=2.24.2",
    "google-genai>=1.15.0",
    "python-frontmatter>=1.1.0",
    "tqdm>=4.67.1",
]
